+++ src/render/processor.rs
@@ -102,22 +102,17 @@ impl<'a> AudioParamValues<'a> {
         Self { nodes }
     }
 
-    pub(crate) fn get_raw(&self, index: &AudioParamId) -> AudioRenderQuantum {
-        self.nodes
-            .get(&index.into())
-            .unwrap()
-            .borrow()
-            .get_buffer()
-            .clone()
-    }
-
     /// Get the computed values for the given [`crate::param::AudioParam`]
     ///
     /// For k-rate params or if the (a-rate) parameter is constant for this block, it will
     /// provide a slice of length 1. In other cases, i.e. a-rate param with scheduled
     /// automations it will provide a slice of length [`crate::RENDER_QUANTUM_SIZE`]
     #[allow(clippy::missing_panics_doc)]
-    pub fn get(&self, index: &AudioParamId) -> impl std::ops::Deref<Target = [f32]> + '_ {
+    pub fn get(&self, index: &AudioParamId) -> impl Deref<Target = [f32]> + '_ {
         DerefAudioRenderQuantumChannel(self.nodes.get(&index.into()).unwrap().borrow())
+    }
+
+    pub(crate) fn listener_params(&self) -> [impl Deref<Target = [f32]> + '_; 9] {
+        crate::context::LISTENER_AUDIO_PARAM_IDS.map(|p| self.get(&p))
     }
 }